Added SamIAm networks that can be re-openned
The exported networks from CaMML can be opened in Netica, but you can't
save anything with over 15 nodes in the Netica limited version. Thus,
follow the README to be able to visualize the data in SamIAm. Or, look
load the .net files I included in the data/ directory for SamIAM.